requirements.txt
# requirements.txt
# Core dependencies for the Task Manager application
# Run with:
#   Bash: PYTHONPATH=. streamlit run src/core/app.py
#   PowerShell: $env:PYTHONPATH = "C:\cookie\task manager"; streamlit run src/core/app.py

streamlit>=1.24.0
crewai>=0.28.8
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.2.0
litellm>=1.35.16
tenacity>=8.2.0